pyproject.toml
[tool.poetry]
name = "tenant-service-api"
version = "0.1.0"
description = ""
authors = ["Camryn Crowder <camryn.crowder@apkudo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
appdirs = "1.4.4"
asgiref = "3.8.1"
asttokens = "3.0.0"
attrs = "24.3.0"
black = "24.10.0"
blinker = "1.4"
boto3 = "1.36.6"
botocore = "1.36.6"
certifi = "2024.12.14"
cffi = "1.17.1"
charset-normalizer = "3.4.1"
click = "8.1.8"
coverage = "7.6.10"
cryptography = "44.0.0"
decorator = "5.1.1"
distro = "1.7.0"
django = "5.1.4"
django-bakery = "0.13.4"
django-cors-headers = "4.6.0"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.4.0"
drf-spectacular = "0.28.0"
environs = "14.1.0"
executing = "2.2.0"
fs = "2.4.16"
httplib2 = "0.20.2"
idna = "3.10"
importlib-metadata = "4.6.4"
inflection = "0.5.1"
iniconfig = "2.0.0"
ipython = "8.31.0"
jedi = "0.19.2"
jmespath = "1.0.1"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
marshmallow = "3.26.0"
matplotlib-inline = "0.1.7"
model-bakery = "1.20.1"
more-itertools = "8.10.0"
mypy-extensions = "1.0.0"
oauthlib = "3.2.0"
packaging = "24.2"
parso = "0.8.4"
pathspec = "0.12.1"
pexpect = "4.9.0"
platformdirs = "4.3.6"
pluggy = "1.5.0"
prompt-toolkit = "3.0.50"
psycopg2-binary = "2.9.10"
pure-eval = "0.2.3"
pycparser = "2.22"
pygments = "2.19.1"
pyjwt = "2.10.1"
pyparsing = "3.1.2"
pytest = "8.3.4"
pytest-cov = "6.0.0"
pytest-django = "4.9.0"
python-dateutil = "2.9.0.post0"
python-decouple = "3.8"
python-dotenv = "1.0.1"
pyyaml = "6.0.2"
referencing = "0.36.1"
requests = "2.32.3"
rpds-py = "0.22.3"
s3transfer = "0.11.2"
setuptools = "75.8.0"
six = "1.16.0"
sqlparse = "0.4.4"
stack-data = "0.6.3"
traitlets = "5.14.3"
typing-extensions = "4.12.2"
uritemplate = "4.1.1"
urllib3 = "2.3.0"
wcwidth = "0.2.13"
zipp = "1.0.0"
kafka-python-ng = "^2.2.3"
drf-nested-routers = "^0.94.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
